refactor: extract model types into types.ts
Move MenuNode, ModelItem, ModelRoot, GetChild, IsLeaf, IfLeaf,
LiteralId, LiteralIds, MarkingMenuItemInput, MarkingMenuInput
from model.ts into types.ts.

Move EmptyTuple, EmptyArray, IsTuple, Point, Segment into utils.ts.
Update imports across all affected files.
