+++ src/model.test.ts
@@ -1,4 +1,5 @@
-import createModel, { type MarkingMenuItemInput } from './model.js';
+import createModel from './model.js';
+import type { MarkingMenuItemInput } from './types.js';
 
 const fourItems = [
   { id: 'right', label: 'Right' },
